Allow empty depends and makedepends fields
The script currently does not correctly see a VITABUILD as having no dependencies if it has ``depends=()`` in it. This makes it so if ``depends=`` is found in the file, but it doesn't contain anything, it will still be seen as a package without dependencies by the automated build system.