Minizip-ng: Fix PIC/MZ_COMPAT
`MZ_COMPAT` was disabled for who-knows-what reason which causes many projects that rely on the simplest unzip functions to throw a linking error. The library was also always being built with PIC resulting in an `Invalid relocation type 25!` error.

This resolves both issues and allows the library to finally be used as intended.